requirements.txt
argon2-cffi==23.1.0
requests==2.32.3
reportlab==4.2.2
rich==13.7.1